[refactor] Stream concerns should be in their own module
