+++ packages/lwdita-xdita/converter.ts
@@ -1,8 +1,9 @@
 import * as saxes from "saxes";
 import { BaseNode, DocumentNode } from "@evolvedbinary/lwdita-ast/nodes";
 import { createNode } from "@evolvedbinary/lwdita-ast/factory";
+import { InMemorySimpleTextStreamCollector } from "./stream";
 import { JDita } from "./classes";
-import { InMemoryTextOutputStream, XditaSerializer } from "./xdita-serializer";
+import { XditaSerializer } from "./xdita-serializer";
 
 /** TODO: Add tests for this module */
 
@@ -128,7 +129,7 @@ export async function xditaToJson(xml: string, abortOnError = true): Promise<JDi
  * @returns The transformed document as an array of XMLTag objects
  */
 export function serializeToXML(root: DocumentNode, indentationChar?: string, tabSize?: number): string {
-  const outStream = new InMemoryTextOutputStream();
+  const outStream = new InMemorySimpleTextStreamCollector();
   let visitor: XditaSerializer;
   if(indentationChar === '\t') {
     visitor = new XditaSerializer(outStream, true, '\t');
